Modernize - should work in py2.7/py3.6/py3.7/py3.8 - use click for a nicer CLI experience - just use RMS from numpy instead of pylab - update gitignore, setup.py
