+++ snowmicrotoolset/utils/__init__.py
@@ -1,3 +1,3 @@
-from smp_optimize import xcorr_microstructure
-from tools_ext import detect_peaks
-from math_stuff import rolling_window, gen_msCoef, xcorr
+from .smp_optimize import xcorr_microstructure
+from .tools_ext import detect_peaks
+from .math_stuff import rolling_window, gen_msCoef, xcorr
