+++ snowmicrotoolset/utils/math_stuff.py
@@ -1,5 +1,5 @@
 import numpy as np
-from pylab import rms_flat
+
 
 def gen_msCoef(dens, ssa):
     '''
@@ -111,7 +111,7 @@ def xcorr(x, y=None, maxlags=None, norm='biased'):
         res = res[lags] / (float(N)-abs(np.arange(-N+1, N)))[lags]
     elif norm == 'coeff':        
         Nf = float(N)
-        rms = rms_flat(x) * rms_flat(y)
+        rms = np.sqrt(np.mean(np.absolute(x)**2)) *  np.sqrt(np.mean(np.absolute(y)**2))
         res = res[lags] / rms / Nf
     else:
         res = res[lags]
